Done with Kanban

+++ client/src/pages/Main.vue
@@ -1,26 +1,31 @@
 <template>
     <div>
         <Navbar @homePage='homeBtn' @addCategory='addForm' @logout='logoutUser'></Navbar>
+        <ErrorMessage v-show="seen" :error='error'></ErrorMessage>
         <div v-if="pageName == 'home'" class="row justify-content-around">
-            <Category 
-                v-for="category in categories" 
-                :key="category.id" 
-                :category="category" 
-                :task="task"
-                :user='users'
-                @addNewTask='addTask'
-                @deleteDetailTask='deleteTask'>
-            </Category>
+                <Category 
+                    v-for="category in categories" 
+                    :key="category.id" 
+                    :category="category" 
+                    :task="task"
+                    :user='users'
+                    @addNewTask='addTask'
+                    @deleteDetailTask='deleteTask'
+                    @editDetailTask='editTask'>
+                </Category>
         </div>
         <AddCategory v-if="pageName == 'addForm'" @addNewCategory="newCategory"></AddCategory>
+        <EditTask v-if="pageName == 'editForm'" @updatedTask='updateTask' :categories='categories' :taskDetail='taskDetail'></EditTask>
     </div>
 </template>
 
 <script>
 import Category from '../components/Category'
 import AddCategory from '../components/CategoryAddForm'
+import EditTask from '../components/EditTaskForm'
 import Navbar from '../components/Navbar'
-import axios from 'axios'
+import ErrorMessage from '../components/ErrorMessage'
+import axios from '../config/axios'
 export default {
     name: 'Main',
     data() {
@@ -28,6 +33,9 @@ export default {
             categories: [],
             task: [],
             users: [],
+            taskDetail: {},
+            error: '',
+            seen: false,
             pageName: 'home'
         }
     },
@@ -35,7 +43,7 @@ export default {
         fetchCategory(){
             axios({
                 method: 'GET',
-                url: 'http://localhost:3000/categories/',
+                url: '/categories/',
                 headers: {
                     access_token: localStorage.access_token
                 }
@@ -50,7 +58,7 @@ export default {
         fetchTask(){
             axios({
                 method: 'GET',
-                url: 'http://localhost:3000/tasks/',
+                url: '/tasks/',
                 headers: {
                     access_token: localStorage.access_token
                 }
@@ -65,7 +73,7 @@ export default {
         fetchUser(){
             axios({
                 method: 'GET',
-                url: 'http://localhost:3000/',
+                url: '/',
                 headers: {
                     access_token: localStorage.access_token
                 }
@@ -80,7 +88,7 @@ export default {
         addTask(payload){
             axios({
                 method: 'POST',
-                url: 'http://localhost:3000/tasks/add',
+                url: '/tasks/add',
                 headers: {
                     access_token: localStorage.access_token
                 },
@@ -90,25 +98,26 @@ export default {
                 }
             })
             .then(data => {
-                console.log(data.data.data)
                 this.task.push(data.data.data)
             })
             .catch(err => {
-                    console.log(err)
+                    this.error = err.response.data.msg
+                    this.seen = true
                 })
         },
         addForm(payload){
             this.pageName = payload
+            this.seen = false
         },
         newCategory(payload){
             axios({
                 method: 'POST',
-                url: 'http://localhost:3000/categories/add',
+                url: '/categories/add',
                 headers: {
                     access_token: localStorage.access_token
                 },
                 data: {
-                    name: payload.title,
+                    name: payload.name
                 }
             })
             .then(({data}) => {
@@ -116,30 +125,81 @@ export default {
                 this.pageName = "home"
             })
             .catch(err => {
-                    console.log(err)
+                console.log(err.response)
+                this.error = err.response.data.msg
+                this.seen = true
             })
         },
         homeBtn(payload){
             this.pageName = payload;
+            this.seen = false
         },
         logoutUser(payload){
             this.$emit('logout', payload)
+            this.seen = false
         },
         deleteTask(payload){
-            axios({
-                method: 'DELETE',
-                url: `http://localhost:3000/tasks/${payload}`,
-                headers: {
-                    access_token: localStorage.access_token
+            bootbox.confirm("Are you sure?", (result) => {
+                console.log(result)
+                if(result){
+                    axios({
+                        method: 'DELETE',
+                        url: `/tasks/${payload}`,
+                        headers: {
+                            access_token: localStorage.access_token
+                        }
+                    })
+                    .then(_=> {
+                        console.log(this.task)
+                        let index = this.task.findIndex(el => el.id == payload)
+                        // for (let i = 0; i < this.task.length; i++){
+                        //     console.log(this.task[i])
+                        //     if(this.task[i].id == payload){
+                        //         index = i
+                        //         break
+                        //     }
+                        // }
+                        
+                        console.log('Berhasil', '====setelah splice')
+                        this.task.splice(index, 1)
+                        this.pageName = 'home'
+                    })
+                    .catch(err => {
+                        console.log('error')
+                            this.error = err.response.data.msg
+                            this.seen = true
+                    })
                 }
             })
-            .then(_=> {
-                let index = this.task.findIndex(el => el.id == payload)
-                this.task.splice(index, 1)
+            
+        }, 
+        editTask(payload){
+            this.pageName = payload.pageName
+            this.taskDetail = payload.data
+            this.seen = false
+        },
+        updateTask(payload){
+            axios({
+                method: 'PUT',
+                url: `/tasks/${payload.id}`,
+                headers: {
+                    access_token: localStorage.access_token
+                },
+                data: {
+                    title: payload.title,
+                    categoryId: payload.selected
+                }
             })
-            .catch(err => {
+                .then(({data}) => {
+                    let index = this.task.findIndex(el => el.id == payload.id)
+                    this.task.splice(index, 1, data.task)
+                    this.pageName = 'home'
+                })
+                .catch(err => {
                     console.log(err)
-            })
+                    this.error = err.response.data.msg
+                    this.seen = true
+                })
         }
     },
     created() {
@@ -150,7 +210,9 @@ export default {
     components: {
         Category,
         Navbar,
-        AddCategory
+        AddCategory,
+        EditTask,
+        ErrorMessage
     }
 }
 </script>
